refactor(tools): split UI asset staging into cohesive files
